Minimum requirements met
At this point, I believe I have all the minimum requirements for this project complete. The working branch will continue on from here, but I'm putting a new master of all the work to date up as a fallback in case something happens. Current build functions (for myself) on IE11 Edge and Firefox web browsers with two major breakpoints built into the code at 478px and 768px. It includes a Readme with instructions for how to run the program, and a short purpose statement, and the code has comments, though I will probably go back soon and write more comments. Much of it is straight forward and to the point. Nothing flashy. Just sufficient at this stage.